--- types/index.d.ts
@@ -1,7 +1,0 @@
-export * from "./items";
-export * from "./type-guards";
-export * from "./models";
-
-export interface JWTinfo {
-  id: string;
-}
